requirements.txt
numpy==1.24
pandas==2.3.1
peft==0.17.1
pytest==8.4.1
joblib==1.5.2
torch==2.8.0
transformers==4.57.0
Flask==3.1.2
Flask-Limiter==4.0.0
email-validator==2.3.0
redis==6.4.0
argon2-cffi==25.1.0
PyJWT==2.10.1